murmurhash

+++ inc/OrdinaryBloomFilter.hpp
@@ -3,6 +3,7 @@
 
 #include <vector>
 #include "AbstractBloomFilter.hpp"
+#include "MurmurHash.hpp"
 
 // forward decl
 namespace bloom {
@@ -28,22 +29,22 @@ namespace bloom {
          *  @see AbstractBloomFilter::AbstractBloomFilter
          */
         explicit
-        OrdinaryBloomFilter(uint8_t numHashes, uint16_t numBits)
+        OrdinaryBloomFilter(uint8_t numHashes, size_t numBits)
                 : AbstractBloomFilter<T>(numHashes, numBits)
         {
             m_bitarray.resize(numBits);
-            for(uint16_t i = 0; i < super::GetNumBits(); i++){
+            for(size_t i = 0; i < super::GetNumBits(); i++){
                 m_bitarray[i] = 0;
             }
 
         }
 
         // Construct from bloom vector
-        OrdinaryBloomFilter(uint8_t numHashes, uint16_t numBits, int* bloom)
+        OrdinaryBloomFilter(uint8_t numHashes, size_t numBits, int* bloom)
                 : AbstractBloomFilter<T>(numHashes, numBits)
         {
             m_bitarray.resize(numBits);
-            for(uint16_t i = 0; i < super::GetNumBits(); i++){
+            for(size_t i = 0; i < super::GetNumBits(); i++){
                 m_bitarray[i] = (bloom[i] == 0) ? 0 : true;
             }
         }
@@ -65,14 +66,14 @@ namespace bloom {
 
         virtual void Serialize(std::ostream &os) const {
             uint8_t numHashes = super::GetNumHashes();
-            uint16_t numBits = super::GetNumBits();
+            size_t numBits = super::GetNumBits();
 
             os.write((const char *) &numHashes, sizeof(uint8_t));
-            os.write((const char *) &numBits, sizeof(uint16_t));
+            os.write((const char *) &numBits, sizeof(size_t));
 
-            for(uint16_t i = 0; i < (numBits + 7) / 8; i++){
+            for(size_t i = 0; i < (numBits + 7) / 8; i++){
                 uint8_t byte = 0;
-                for(int j = 0; j < 8 && (i + j) < numBits; j++){
+                for(size_t j = 0; j < 8 && (i + j) < numBits; j++){
                     byte = (byte << 1) | m_bitarray[8*i+j];
                 }
                 os.write((const char *) &byte, sizeof(uint8_t));
@@ -91,17 +92,17 @@ namespace bloom {
          */
         static OrdinaryBloomFilter<T> Deserialize(std::istream &is){
             uint8_t numHashes;
-            uint16_t numBits;
+            size_t numBits;
 
             is.read((char *) &numHashes, sizeof(uint8_t));
-            is.read((char *) &numBits, sizeof(uint16_t));
+            is.read((char *) &numBits, sizeof(size_t));
 
             OrdinaryBloomFilter<T> r (numHashes, numBits);
 
-            for(uint16_t i = 0; i < (numBits + 7) / 8; i++){
+            for(size_t i = 0; i < (numBits + 7) / 8; i++){
                 uint8_t byte;
                 is.read((char *) &byte, sizeof(uint8_t));
-                for(int j = 0; j < 8 && (i + j) < numBits; j++){
+                for(size_t j = 0; j < 8 && (i + j) < numBits; j++){
                     r.m_bitarray[8 * i + j] = byte & (1 << (7-j));
                 }
             }
@@ -115,12 +116,12 @@ namespace bloom {
          *  @return A new OrdinaryBloomFilter with half as many bits.
          */
         OrdinaryBloomFilter<T> Compress() const {
-            uint16_t oldNumBits = super::GetNumBits();
-            uint16_t newNumBits = oldNumBits / 2;
+            size_t oldNumBits = super::GetNumBits();
+            size_t newNumBits = oldNumBits / 2;
 
             OrdinaryBloomFilter<T> res(super::GetNumHashes(), newNumBits);
 
-            for(unsigned i = 0; i < oldNumBits; i++){
+            for(size_t i = 0; i < oldNumBits; i++){
                 res.m_bitarray[i % newNumBits] = res.m_bitarray[i % newNumBits] | m_bitarray[i];
             }
 
@@ -133,9 +134,9 @@ namespace bloom {
          *  @return A new PairedBloomFilter
          */
         PairedBloomFilter<T> ToPairedBloomFilter() const {
-            uint16_t numBits = super::GetNumBits();
+            size_t numBits = super::GetNumBits();
             PairedBloomFilter<T> res(super::GetNumHashes(), numBits);
-            for(unsigned i = 0; i < numBits; i++){
+            for(size_t i = 0; i < numBits; i++){
                 res.m_bitarray[i] = m_bitarray[i];
             }
             return res;
@@ -148,7 +149,7 @@ namespace bloom {
          *  @param other BF to combine into this one
          */
         void Union(OrdinaryBloomFilter<T> const& other){
-            for(unsigned i = 0; i < super::GetNumBits(); i++){
+            for(size_t i = 0; i < super::GetNumBits(); i++){
                 m_bitarray[i] = m_bitarray[i] | other.m_bitarray[i];
             }
         }
